pyproject.toml
[tool.poetry]
name = "openapi-sorter"
version = "0.1.0"
description = "OpenAPI-Sorter is a utility for parsing, sorting, and outputting OpenAPI YAML files by organizing path, model, and other notations. It helps maintain clean and well-structured API documentation by ensuring consistent ordering of components in the YAML file."
authors = ["Sugiyama Tsuyoshi <sugiyama@fullmarks.co.jp>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
argparse = "^1.4.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 120
balanced_wrapping = true
multi_line_output = 3
include_trailing_comma = true

[tool.black]
line-length = 120
skip-string-normalization = true


[tool.poetry.scripts]
openapi-sorter = 'openapi_sorter.cli:main'